move functions to proper file
